+++ src/enemies/enemiesModules.ts
@@ -1,15 +1,12 @@
-function displayData(data: string){
-    let res = data + ' My World 111';
-    return res
-}
-function placeEnemye(){
+
+function placeEnemyes(){
 
 }
 function moveEnemyes(){
-    
+
 }
 function loadEnemyes(){
 
 }
 
-module.exports.enemies = displayData;
+module.exports.enemies = placeEnemyes;
